+++ packages/eslint-config/rules/deprecated/node.js
@@ -1,4 +1,5 @@
 'use strict';
+
 /**
  * The following eslint-plugin-node rules have been deprecated.
  *
@@ -6,7 +7,7 @@
 
 module.exports = {
   rules: {
-    'node/no-unsupported-features': 'off', // node - v7.0.0 - node/no-unsupported-features/es-syntax & node/no-unsupported-features/es-builtins
-    'node/no-hide-core-modules': 'off', // node - v4.2.0
+    'node/no-unsupported-features': 'off', // Node - v7.0.0 - node/no-unsupported-features/es-syntax & node/no-unsupported-features/es-builtins
+    'node/no-hide-core-modules': 'off', // Node - v4.2.0
   },
 };