+++ packages/eslint-config/CHANGELOG.md
@@ -26,6 +26,7 @@ _This sections lists changes committed to master since most recent release_
 - Updated dependencies, including ESLint to v7.0.0
 - Set `react/prop-types` to `off` in TypeScript overrides file, previously set to `error`
 - Deprecated the Node ESLint rules, as they are have been supplanted with eslint-plugin-node
+- Move `id-blacklist` ESLint rule to the deprecated list
 - Move from package from it's own repository into the CryptoPapers Design System monorepo
 - Run tests using Jest from the monorepo root